requirements.txt
pydantic
dash
requests
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pandas
plotly[express]
